# openapi.yaml
openapi: 3.0.3
info:
  title: cerbix API
  version: 0.19.0
  description: >
    Internal uptime & SLA monitoring API. Authentication is via a server-side
    session cookie established by an OIDC provider (`/auth/login` → `/auth/callback`)
    or local login (`/auth/local/login`). All `/api/v1/*` endpoints require a
    valid session and enforce tenant isolation: callers only see and mutate the
    organizations/projects they are authorized for (non-members receive 404;
    insufficient in-org role receives 403).
servers:
  - url: /
    description: Same-origin (SPA is served by the API binary)
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /api/v1/organizations/{orgID}/status-pages:
    parameters: [ { $ref: '#/components/parameters/OrgID' } ]
    get:
      tags: [status-pages]
      summary: List an organization's status pages (org member)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/StatusPage' } }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [status-pages]
      summary: Create a status page (org admin). Unlisted pages get a generated token.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateStatusPage' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatusPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/status-pages/{pageID}:
    parameters: [ { $ref: '#/components/parameters/PageID' } ]
    get:
      tags: [status-pages]
      summary: Get a status page's configuration (org member)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatusPage' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [status-pages]
      summary: Update a status page's title/visibility (org admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateStatusPage' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatusPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [status-pages]
      summary: Delete a status page and its components (org admin)
      responses:
        '204': { description: No Content }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/public/status-pages/{slug}:
    parameters:
      - in: path
        name: slug
        required: true
        schema: { type: string }
      - in: query
        name: token
        required: false
        schema: { type: string }
        description: Required for unlisted pages.
    get:
      tags: [status-pages]
      security: []
      summary: >
        Public (unauthenticated) render. Public pages are served to anyone;
        unlisted pages require the matching token; internal pages return 404.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatusPageRender' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cerbix_session
    bearerAuth:
      type: http
      scheme: bearer
      description: "Machine auth via Authorization: Bearer. Either a cerbix service-account token (cbx_...) or an OIDC client-credentials access token (a realm-issued JWT, when OIDC is enabled); the JWT subject is JIT-provisioned and access is membership-gated."
  parameters:
    OrgID:
      in: path
      name: orgID
      required: true
      schema: { type: string, format: uuid }
...
    PageID:
      in: path
      name: pageID
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Forbidden:
      description: Authenticated but insufficient role.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found or not visible to the caller (existence hidden).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    MaintenanceWindow:
      type: object
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        monitor_id: { type: string, format: uuid, nullable: true }
        starts_at: { type: string, format: date-time }
        ends_at: { type: string, format: date-time }
        reason: { type: string }
        created_at: { type: string, format: date-time }
...
    Incident:
      type: object
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        monitor_id: { type: string, format: uuid, description: 'Set for auto-incidents opened from a monitor going down.' }
        external_key: { type: string, description: 'Correlation key for externally-sourced incidents (e.g. an Alertmanager fingerprint).' }
        title: { type: string }
        status: { $ref: '#/components/schemas/IncidentStatus' }
        impact: { $ref: '#/components/schemas/IncidentImpact' }
        source: { type: string, enum: [manual, api, auto] }
        started_at: { type: string, format: date-time }
        resolved_at: { type: string, format: date-time, nullable: true }
        acknowledged_at: { type: string, format: date-time, nullable: true, description: 'Set when acknowledged; stops on-call escalation.' }
        acknowledged_by: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    IncidentStatus:
      type: string
      enum: [investigating, identified, monitoring, resolved]
    IncidentImpact:
      type: string
      enum: [none, minor, major, critical]
...
    IncidentUpdate:
      type: object
      properties:
        id: { type: string, format: uuid }
        incident_id: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/IncidentStatus' }
        body: { type: string }
        author: { type: string }
        created_at: { type: string, format: date-time }
...
    Postmortem:
      type: object
      properties:
        id: { type: string, format: uuid }
        incident_id: { type: string, format: uuid }
        body: { type: string }
        author: { type: string }
        published_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    IncidentDetail:
      description: A resolved incident with its timeline and (optional) postmortem, for the status page's past-incidents accordion.
      allOf:
        - $ref: '#/components/schemas/Incident'
        - type: object
          properties:
            updates: { type: array, items: { $ref: '#/components/schemas/IncidentUpdate' } }
            postmortem: { $ref: '#/components/schemas/Postmortem' }
...
    Visibility:
      type: string
      enum: [public, internal, unlisted]
    ComponentStatus:
      type: string
      enum: [operational, degraded, partial_outage, major_outage, maintenance]
    StatusPage:
      type: object
      properties:
        id: { type: string, format: uuid }
        org_id: { type: string, format: uuid }
        project_id: { type: string, format: uuid, nullable: true }
        slug: { type: string }
        title: { type: string }
        visibility: { $ref: '#/components/schemas/Visibility' }
        unlisted_token: { type: string, description: 'Present only for unlisted pages.' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateStatusPage:
      type: object
      required: [slug, title]
      properties:
        slug: { type: string }
        title: { type: string }
        visibility: { $ref: '#/components/schemas/Visibility' }
        project_id: { type: string, format: uuid, description: 'Omit for an org-level page.' }
    UpdateStatusPage:
      type: object
      description: 'Partial update; slug and org are immutable. Switching to unlisted mints a token, switching away clears it.'
      properties:
        title: { type: string }
        visibility: { $ref: '#/components/schemas/Visibility' }
...
    ComponentDay:
      type: object
      description: One day of a component's 90-day availability strip.
      properties:
        day: { type: string, description: 'YYYY-MM-DD (UTC)' }
        uptime_percent: { type: number, format: double }
        total: { type: integer, format: int64 }
    ComponentView:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        group: { type: string }
        status: { $ref: '#/components/schemas/ComponentStatus' }
        uptime_90d: { type: number, format: double, nullable: true }
        daily: { type: array, items: { $ref: '#/components/schemas/ComponentDay' }, description: 'Per-day availability; empty for manual components.' }
    StatusPageRender:
      type: object
      properties:
        slug: { type: string }
        title: { type: string }
        visibility: { $ref: '#/components/schemas/Visibility' }
        summary: { $ref: '#/components/schemas/ComponentStatus' }
        components: { type: array, items: { $ref: '#/components/schemas/ComponentView' } }
        active_incidents: { type: array, items: { $ref: '#/components/schemas/IncidentDetail' } }
        recent_incidents: { type: array, items: { $ref: '#/components/schemas/IncidentDetail' }, description: 'Resolved in the last 90 days, newest first; each carries its timeline and postmortem (if published).' }
        maintenance: { type: array, items: { $ref: '#/components/schemas/MaintenanceWindow' }, description: 'Active or scheduled maintenance.' }
        updated_at: { type: string, format: date-time }
...